+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * Added "Set as Target" in Monitor. (HaullyGames)
 * Panel - Define the hover area to be the same size of font size. (HaullyGames)
 * Increased SCANSat experiment sizes. A planetary map shouldn't be smaller than a temperature reading. (Sir Mortimer)
+* Connection Manager Panel, click in signal Icon to open the Connection Panel. (HaullyGames)
+* CommNet: disable stock transmission buttons and transmission action group. (HaullyGames)
 
 ------------------------------------------------------------------------------------------------------